requirements.txt
Django==4.2.10
djangorestframework==3.14.0
django-cors-headers==4.3.1
mysqlclient==2.2.7
drf-yasg==1.21.7
django-filter==25.1
Pillow==10.2.0
